made code more modular by dividing up the rendering for the specific listing page
